Split runtime metadata parsing by program format and architecture
