+++ src/runtime_metadata/loader.rs
@@ -1,0 +1,57 @@
+use std::{io, str};
+
+use bad64::DecodeError;
+use thiserror::Error;
+
+#[cfg(feature = "elf")]
+pub mod elf;
+
+#[derive(Error, Debug)]
+pub enum Il2CppBinaryError {
+    #[error("error disassembling code")]
+    Disassemble(DecodeError),
+
+    #[error("failed to convert virtual address {0:#016x}")]
+    VAddrConv(u64),
+
+    #[error("could not find il2cpp_init symbol in elf")]
+    MissingIl2CppInit,
+
+    #[error("could not find registration function")]
+    MissingRegistration,
+
+    #[error("invalid Il2CppType with type {0}")]
+    InvalidType(u8),
+
+    #[error(transparent)]
+    Io(#[from] io::Error),
+
+    #[error(transparent)]
+    BinaryDeserialize(#[from] binread::Error),
+
+    #[error(transparent)]
+    Utf8(#[from] str::Utf8Error),
+
+    #[error(transparent)]
+    Elf(#[from] object::Error),
+}
+
+pub type Result<T> = std::result::Result<T, Il2CppBinaryError>;
+
+#[derive(Error, Debug, Clone, Copy)]
+#[error("error disassembling code")]
+pub struct DisassembleError;
+
+pub fn strlen(data: &[u8], offset: usize) -> usize {
+    let mut len = 0;
+    while data[offset + len] != 0 {
+        len += 1;
+    }
+    len
+}
+
+pub fn get_str(data: &[u8], offset: usize) -> Result<&str> {
+    let len = strlen(data, offset);
+    let str = str::from_utf8(&data[offset..offset + len])?;
+    Ok(str)
+}
